% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSMining.R
\docType{data}
\name{BuildOperation}
\alias{BuildOperation}
\title{An example 1-week data set containing the power consumption data of two building services sub-systems}
\format{A data frame with 672 rows and 6 variables}
\usage{
data(BuildOperation)
}
\description{
The data contain 7-day power consumption data of two building services sub-systems, i.e., water-cooled chillers (WCC) and air handling unit (AHU)
The collection interval is 15-minute. The data have 672 observations and 6 variables, i.e., Month, Day, Hour, Minute, WCC and AHU
}
\examples{
library(ggplot2)
data(BuildOperation)
ggplot(data = BuildOperation, aes(x = 1:dim(BuildOperation)[1], y = WCC)) +
 geom_line() + geom_point()
ggplot(data = BuildOperation, aes(x = 1:dim(BuildOperation)[1], y = AHU)) +
 geom_line() + geom_point()
}
\keyword{datasets}

