% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSMining.R
\name{Func.matrix}
\alias{Func.matrix}
\title{A function to create the distance matrix for alphabets}
\usage{
Func.matrix(a)
}
\arguments{
\item{a}{is an integer specifying the alphabet size.}
}
\value{
The function returns a matrix showing the distance between alphabets
}
\description{
This function create a distance matrix for alphabets used for SAX transformation
}
\examples{
Func.matrix(a=5)
}

