\name{TSMining-package}
\alias{TSMining-package}
\alias{TSMining}
\docType{package}
\title{
Implementation of Univariate and Multivariate Motif Discovery in Time-Series Data
}
\description{
Implementations of a number of functions used to mine numeric time-series data. It covers the implementation of SAX transformation, univariate motif discovery (based on the random projection method), multivariate motif discovery (based on graph clustering), and several functions used for the ease of visualizing the motifs discovered. The details of SAX transformation can be found in J. Lin. E. Keogh, L. Wei, S. Lonardi, Experiencing SAX: A novel symbolic representation of time series, Data Mining and Knowledge Discovery 15 (2) (2007) 107-144. Details on univariate motif discovery method implemented can be found in B. Chiu, E. Keogh, S. Lonardi, Probabilistic discovery of time series motifs, ACM SIGKDD, Washington, DC, USA, 2003, pp. 493-498. Details on the multivariate motif discovery method implemented can be found in A. Vahdatpour, N. Amini, M. Sarrafzadeh, Towards unsupervised activity discovery using multi-dimensional motif detection in time series, IJCAI 2009 21st International Joint Conference on Artificial Intelligence.
}
\details{
\tabular{ll}{
Package: \tab TSMining\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-24\cr
License: \tab GPL-3 \cr
}

}
\author{
Cheng Fan

Maintainer: Cheng Fan <raja8885@hotmail.com>
Cheng Fan
}
\references{
J. Lin. E. Keogh, L. Wei, S. Lonardi. Experiencing SAX: A novel symbolic representation of time series. Data Mining and Knowledge Discovery 15 (2) (2007) 107-144. DOI:http://cs.gmu.edu/~jessica/SAX_DAMI_preprint.pdf

B. Chiu, E. Keogh, S. Lonardi. Probabilistic discovery of time series motifs. ACM SIGKDD, Washington, DC, USA, 2003, pp. 493-498. DOI:10.1145/956750.956808

A. Vahdatpour, N. Amini, M. Sarrafzadeh. Towards unsupervised activity discovery using multi-dimensional motif detection in time series. IJCAI 2009 21st International Joint Conference on Artificial Intelligence. DOI:http://ijcai.org/papers09/Papers/IJCAI09-212.pdf

J. Buhler, M. Tompa. Finding motifs using random projections. 5th Annual International Conference on Computational Molecular Biology, ACM New York, New York, USA, 2001, pp. 69-76. DOI:http://www.cs.columbia.edu/~cleslie/cs4761/papers/buhler_finding.pdf
}
